+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Week recurrence for "do" command.
+
+## [0.5.0] - 2024-04-03
+
+### Changed
+- Remove some code to simplify program logic.
+
 ## [0.4.0] - 2023-11-26
 
 ### Added